+++ packages/astro/src/core/preview/index.ts
@@ -3,6 +3,7 @@ import type { AddressInfo } from 'net';
 import type { AstroConfig } from '../../@types/astro';
 import type { LogOptions } from '../logger/core';
 
+import fs from 'fs';
 import http from 'http';
 import { performance } from 'perf_hooks';
 import sirv from 'sirv';
@@ -77,7 +78,18 @@ export default async function preview(
 			default: {
 				// HACK: rewrite req.url so that sirv finds the file
 				req.url = '/' + req.url?.replace(baseURL.pathname, '');
-				staticFileServer(req, res, () => sendError('Not Found'));
+				staticFileServer(req, res, () => {
+					const errorPagePath = fileURLToPath(config.outDir + '/404.html');
+					if (fs.existsSync(errorPagePath)) {
+						res.statusCode = 404;
+						res.setHeader('Content-Type', 'text/html;charset=utf-8');
+						res.end(fs.readFileSync(errorPagePath));
+					} else {
+						staticFileServer(req, res, () => {
+							sendError('Not Found');
+						});
+					}
+				});
 				return;
 			}
 		}