+++ packages/integrations/mdx/src/index.ts
@@ -1,40 +1,73 @@
 import { compile as mdxCompile, nodeTypes } from '@mdx-js/mdx';
 import mdxPlugin, { Options as MdxRollupPluginOptions } from '@mdx-js/rollup';
-import type { AstroIntegration } from 'astro';
+import type { AstroConfig, AstroIntegration } from 'astro';
 import { parse as parseESM } from 'es-module-lexer';
 import rehypeRaw from 'rehype-raw';
-import remarkFrontmatter from 'remark-frontmatter';
 import remarkGfm from 'remark-gfm';
-import type { RemarkMdxFrontmatterOptions } from 'remark-mdx-frontmatter';
-import remarkMdxFrontmatter from 'remark-mdx-frontmatter';
 import remarkShikiTwoslash from 'remark-shiki-twoslash';
 import remarkSmartypants from 'remark-smartypants';
 import { VFile } from 'vfile';
 import type { Plugin as VitePlugin } from 'vite';
+import { rehypeApplyFrontmatterExport, remarkInitializeAstroData } from './astro-data-utils.js';
 import rehypeCollectHeadings from './rehype-collect-headings.js';
 import remarkPrism from './remark-prism.js';
-import { getFileInfo, getFrontmatter } from './utils.js';
+import { getFileInfo, parseFrontmatter } from './utils.js';
 
 type WithExtends<T> = T | { extends: T };
 
 type MdxOptions = {
 	remarkPlugins?: WithExtends<MdxRollupPluginOptions['remarkPlugins']>;
 	rehypePlugins?: WithExtends<MdxRollupPluginOptions['rehypePlugins']>;
-	/**
-	 * Configure the remark-mdx-frontmatter plugin
-	 * @see https://github.com/remcohaszing/remark-mdx-frontmatter#options for a full list of options
-	 * @default {{ name: 'frontmatter' }}
-	 */
-	frontmatterOptions?: RemarkMdxFrontmatterOptions;
 };
 
-const DEFAULT_REMARK_PLUGINS = [remarkGfm, remarkSmartypants];
-const DEFAULT_REHYPE_PLUGINS = [rehypeCollectHeadings];
+const DEFAULT_REMARK_PLUGINS: MdxRollupPluginOptions['remarkPlugins'] = [
+	remarkGfm,
+	remarkSmartypants,
+];
+const DEFAULT_REHYPE_PLUGINS: MdxRollupPluginOptions['rehypePlugins'] = [];
 
 function handleExtends<T>(config: WithExtends<T[] | undefined>, defaults: T[] = []): T[] {
 	if (Array.isArray(config)) return config;
 
 	return [...defaults, ...(config?.extends ?? [])];
+}
+
+function getRemarkPlugins(
+	mdxOptions: MdxOptions,
+	config: AstroConfig
+): MdxRollupPluginOptions['remarkPlugins'] {
+	let remarkPlugins = [
+		// Initialize vfile.data.astroExports before all plugins are run
+		remarkInitializeAstroData,
+		...handleExtends(mdxOptions.remarkPlugins, DEFAULT_REMARK_PLUGINS),
+	];
+	if (config.markdown.syntaxHighlight === 'shiki') {
+		// Default export still requires ".default" chaining for some reason
+		// Workarounds tried:
+		// - "import * as remarkShikiTwoslash"
+		// - "import { default as remarkShikiTwoslash }"
+		const shikiTwoslash = (remarkShikiTwoslash as any).default ?? remarkShikiTwoslash;
+		remarkPlugins.push([shikiTwoslash, config.markdown.shikiConfig]);
+	}
+	if (config.markdown.syntaxHighlight === 'prism') {
+		remarkPlugins.push(remarkPrism);
+	}
+	return remarkPlugins;
+}
+
+function getRehypePlugins(
+	mdxOptions: MdxOptions,
+	config: AstroConfig
+): MdxRollupPluginOptions['rehypePlugins'] {
+	let rehypePlugins = handleExtends(mdxOptions.rehypePlugins, DEFAULT_REHYPE_PLUGINS);
+
+	if (config.markdown.syntaxHighlight === 'shiki' || config.markdown.syntaxHighlight === 'prism') {
+		rehypePlugins.push([rehypeRaw, { passThrough: nodeTypes }]);
+	}
+	// getHeadings() is guaranteed by TS, so we can't allow user to override
+	rehypePlugins.push(rehypeCollectHeadings);
+
+	return rehypePlugins;
 }
 
 export default function mdx(mdxOptions: MdxOptions = {}): AstroIntegration {
@@ -43,38 +76,10 @@ export default function mdx(mdxOptions: MdxOptions = {}): AstroIntegration {
 		hooks: {
 			'astro:config:setup': ({ updateConfig, config, addPageExtension, command }: any) => {
 				addPageExtension('.mdx');
-				let remarkPlugins = handleExtends(mdxOptions.remarkPlugins, DEFAULT_REMARK_PLUGINS);
-				let rehypePlugins = handleExtends(mdxOptions.rehypePlugins, DEFAULT_REHYPE_PLUGINS);
-
-				if (config.markdown.syntaxHighlight === 'shiki') {
-					remarkPlugins.push([
-						// Default export still requires ".default" chaining for some reason
-						// Workarounds tried:
-						// - "import * as remarkShikiTwoslash"
-						// - "import { default as remarkShikiTwoslash }"
-						(remarkShikiTwoslash as any).default ?? remarkShikiTwoslash,
-						config.markdown.shikiConfig,
-					]);
-					rehypePlugins.push([rehypeRaw, { passThrough: nodeTypes }]);
-				}
-
-				if (config.markdown.syntaxHighlight === 'prism') {
-					remarkPlugins.push(remarkPrism);
-					rehypePlugins.push([rehypeRaw, { passThrough: nodeTypes }]);
-				}
-
-				remarkPlugins.push(remarkFrontmatter);
-				remarkPlugins.push([
-					remarkMdxFrontmatter,
-					{
-						name: 'frontmatter',
-						...mdxOptions.frontmatterOptions,
-					},
-				]);
 
 				const mdxPluginOpts: MdxRollupPluginOptions = {
-					remarkPlugins,
-					rehypePlugins,
+					remarkPlugins: getRemarkPlugins(mdxOptions, config),
+					rehypePlugins: getRehypePlugins(mdxOptions, config),
 					jsx: true,
 					jsxImportSource: 'astro',
 					// Note: disable `.md` support
@@ -93,24 +98,23 @@ export default function mdx(mdxOptions: MdxOptions = {}): AstroIntegration {
 								async transform(code, id) {
 									if (!id.endsWith('mdx')) return;
 
-									// If user overrides our default YAML parser,
-									// do not attempt to parse the `layout` via gray-matter
-									if (!mdxOptions.frontmatterOptions?.parsers) {
-										const frontmatter = getFrontmatter(code, id);
-										if (frontmatter.layout) {
-											const { layout, ...content } = frontmatter;
-											code += `\nexport default async function({ children }) {\nconst Layout = (await import(${JSON.stringify(
-												frontmatter.layout
-											)})).default;\nconst frontmatter=${JSON.stringify(
-												content
-											)};\nreturn <Layout frontmatter={frontmatter} content={frontmatter} headings={getHeadings()}>{children}</Layout> }`;
-										}
+									let { data: frontmatter, content: pageContent } = parseFrontmatter(code, id);
+									if (frontmatter.layout) {
+										const { layout, ...contentProp } = frontmatter;
+										pageContent += `\n\nexport default async function({ children }) {\nconst Layout = (await import(${JSON.stringify(
+											frontmatter.layout
+										)})).default;\nconst frontmatter=${JSON.stringify(
+											contentProp
+										)};\nreturn <Layout frontmatter={frontmatter} content={frontmatter} headings={getHeadings()}>{children}</Layout> }`;
 									}
 
-									const compiled = await mdxCompile(
-										new VFile({ value: code, path: id }),
-										mdxPluginOpts
-									);
+									const compiled = await mdxCompile(new VFile({ value: pageContent, path: id }), {
+										...mdxPluginOpts,
+										rehypePlugins: [
+											...(mdxPluginOpts.rehypePlugins ?? []),
+											() => rehypeApplyFrontmatterExport(frontmatter),
+										],
+									});
 
 									return {
 										code: String(compiled.value),