+++ packages/integrations/node/CHANGELOG.md
@@ -38,7 +38,7 @@
   The new `Astro.clientAddress` property allows you to get the IP address of the requested user.
 
   ```astro
-  <div>Your address { Astro.clientAddress }</div>
+  <div>Your address {Astro.clientAddress}</div>
   ```
 
   This property is only available when building for SSR, and only if the adapter you are using supports providing the IP address. If you attempt to access the property in a SSG app it will throw an error.